+++ app/index.js
@@ -1,0 +1,2 @@
+import './api'
+import './router'
