+++ babel.js
@@ -1,3 +1,3 @@
 require('babel-register')
 require('babel-polyfill')
-require('./app')
+require('./app/index')
