+++ lib/handlebars.js
@@ -1,5 +1,5 @@
 import Handlebars from 'handlebars'
-import { domain } from '../data/seed.js'
+import { domain } from '../data/seed'
 
 Handlebars.registerHelper('__appDomain', function() {
   return domain;
