+++ lib/parse.js
@@ -1,4 +1,12 @@
-import chrono from 'chrono-node'
+import chrono from 'chrono-node' // Sugar.js parses some things better
+
+/* 
+let test
+//test = chrono.parseDate("foo the bar by sep 15th noon")
+test = chrono.parseDate("Call_the_dentist by 2017-10-15 12pm")
+console.log(`DEBUG: ${test} (unixtime = ${test.getTime()/1000}`)
+console.log("DEBUG: ", JSON.stringify(parsePromise(promises[2])))
+*/
 
 // Take urtext and return a promise data structure
 export default function parseprom(urtx) {
