+++ styles/_card.scss
