+++ styles/_user.scss
@@ -1,4 +1,4 @@
-.user-promises-list {
+s.user-promises-list {
   h2 {
     background-color: $blue;
     margin: -1rem -1rem 1rem;
