+++ tmp/style.css
@@ -1,0 +1,194 @@
+* {
+  box-sizing: border-box; }
+
+body {
+  font-family: helvetica, arial, sans-serif;
+  padding: 0;
+  margin: 0;
+  background-color: #313131; }
+
+h1, .bold {
+  font-weight: bold; }
+
+h1, h2, h3, h4, h5, p {
+  color: white; }
+
+a {
+  color: #2fa2fa;
+  text-decoration: none; }
+
+form {
+  margin-bottom: 25px;
+  padding: 15px;
+  background-color: cyan;
+  display: inline-block;
+  width: 100%;
+  max-width: 340px;
+  border-radius: 3px; }
+
+input {
+  display: block;
+  margin-bottom: 10px;
+  padding: 5px;
+  width: 100%;
+  border: 1px solid lightgrey;
+  border-radius: 3px;
+  font-size: 16px; }
+
+.blue-border {
+  border: 1px solid #2fa2fa;
+  padding: 1rem; }
+
+.promise-card {
+  background-color: #2fa2fa;
+  border-radius: .25rem;
+  border: 1px groove #7bc9ff;
+  padding: 1rem;
+  transition: background .25s ease-in-out; }
+  .promise-card:hover {
+    background-color: #126bbd; }
+  .promise-card, .promise-card a, .promise-card span {
+    color: white; }
+  .promise-card span {
+    font-weight: 300;
+    font-size: 50%; }
+
+.promise-user {
+  align-items: center;
+  display: flex;
+  margin: 0; }
+  .promise-user > :first-child {
+    display: block;
+    margin-right: auto; }
+
+.promise-domain {
+  font-style: italic;
+  font-size: 75%; }
+
+.promise-text {
+  margin-top: .25rem; }
+
+.promise-action-bar {
+  display: flex;
+  margin: 1rem calc(-1rem - 1px) -1rem; }
+  .promise-action-bar .action-bar-button {
+    flex-basis: 50%; }
+    .promise-action-bar .action-bar-button .promise-action {
+      background-color: #FEFEFE;
+      color: #2fa2fa;
+      display: block;
+      font-size: .75rem;
+      font-weight: bold;
+      margin: 1px;
+      opacity: .5;
+      padding: .5rem;
+      text-align: center;
+      text-transform: uppercase; }
+      .promise-action-bar .action-bar-button .promise-action:hover {
+        opacity: .75; }
+
+.create-promise {
+  text-align: center;
+  padding: 4rem; }
+  .create-promise .promise-statement {
+    margin: 0 0 3rem;
+    line-height: 3rem;
+    font-weight: 300; }
+    .create-promise .promise-statement span {
+      font-weight: bold; }
+  .create-promise .promise-button {
+    border: 1px solid #2fa2fa;
+    border-radius: .5rem;
+    color: #2fa2fa;
+    display: block;
+    font-size: 2rem;
+    font-weight: bold;
+    text-transform: uppercase;
+    padding: 2rem;
+    text-align: center;
+    transition: all .75s ease-in-out; }
+    .create-promise .promise-button:hover {
+      background-color: #2fa2fa;
+      color: white; }
+
+.promises-preview {
+  display: flex;
+  flex-wrap: wrap;
+  margin: .5rem -.5rem 0; }
+  .promises-preview .promise-card {
+    flex-grow: 1;
+    margin: .5rem; }
+
+.user-promises-list h2 {
+  background-color: #2fa2fa;
+  margin: -1rem -1rem 1rem;
+  padding: 1rem; }
+  .user-promises-list h2 a {
+    color: white; }
+
+.user-promises-list .promise-text {
+  margin: 0; }
+  .user-promises-list .promise-text a {
+    margin: -1rem;
+    padding: 1rem; }
+
+.user-promises-list .promise-user, .user-promises-list .promise-user a {
+  display: none; }
+
+.promises-list {
+  display: flex;
+  flex-wrap: wrap;
+  list-style-type: none;
+  margin: 0;
+  padding: 0; }
+  .promises-list > li {
+    margin: .5rem;
+    padding: 0; }
+
+header {
+  background-color: #242424;
+  border-bottom: 1px ridge #424242;
+  display: flex;
+  align-items: center;
+  padding: .5rem 2rem; }
+  header h1 {
+    margin: 0 auto 0 0; }
+  header a.button {
+    float: right;
+    color: #FEFEFE;
+    margin-left: .75rem;
+    padding: .4rem 1rem;
+    font-size: .75rem;
+    border-radius: 3px;
+    background-color: #454545;
+    border: 1px solid #73706d;
+    box-shadow: 2px 2px #313131;
+    cursor: pointer; }
+    header a.button:hover {
+      background-color: #2fa2fa; }
+    header a.button:active {
+      box-shadow: none; }
+
+footer {
+  color: #FEFEFE;
+  margin-top: 1rem;
+  padding: 2rem;
+  border-top: 1px solid #73706d; }
+  footer > a {
+    color: #BBBBBB; }
+
+main {
+  padding: 2rem; }
+
+.sub-heading {
+  margin-top: 0; }
+
+section.single-promise {
+  margin-top: 2rem; }
+  section.single-promise h5 {
+    text-align: right;
+    font-weight: lighter;
+    color: #999; }
+
+.nicejob {
+  text-decoration: line-through; }
