+++ views/layouts/main.handlebars
@@ -10,7 +10,8 @@
 <meta charset="utf-8">
 <meta http-equiv="X-UA-Compatible" content="IE=edge">
 <meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="/style.css">
+<link rel="stylesheet" href="/main.css">
+<script src="/js/promise.js"></script>
 </head>
 <body>
   {{> header}}
